hw/npu: make the Ethos-U executor compute uint8 networks correctly
The executor was int8-only. Running a real uint8 (TFLite "legacy"
quantized) MobileNet-v1 through the eIQ Ethos-U delegate produced the
wrong classification: every conv ran, but the result was garbage. Two
distinct bugs, both specific to the uint8 quantization scheme:

1. Activation signedness was ignored. NPU_SET_IFM/OFM_PRECISION bit 0
   selects signed (int8) vs unsigned (uint8) activations (confirmed
   against Vela's register_command_stream_generator: prec |= 1 if the
   dtype is signed; bit 6 is the NHCWB16 layout, not signedness). The
   decoder only read the bit-depth field, so uint8 feature maps (zero
   point 128) were misinterpreted as int8.

   Decode op->ifm_unsigned/ofm_unsigned and, in exec_conv/exec_pool,
   rebias unsigned activations to int8 (XOR 0x80) on load and back on
   store, shifting ifm_zp/ofm_zp/act_min/act_max by -128. This reuses
   the existing, bit-exact int8 kernels; the int8 path is unchanged
   (the offsets are zero).

2. Weights were truncated to int8. Vela recentres a uint8 weight
   tensor by its (non-128) zero point, so the symmetric weights span a
   9-bit-signed range (e.g. [-151, 104] for weight zero point 151).
   ethos_u_load_weights cast the decoded values to int8, wrapping any
   weight below -128 (e.g. -150 -> +106), corrupting one output
   channel of the first conv and propagating to the classifier.

   Keep weights as int16 from decode through the conv/depthwise
   kernels. int8 models are unaffected (symmetric, in [-127, 127]).

With both fixes every layer of uint8 MobileNet-v1 - all convolutions,
the depthwise stages, the global average pool and the 1x1 classifier -
is bit-exact against the TensorFlow Lite reference, and the NPU logits
argmax matches the host golden. (The trailing softmax is lowered to
elementwise + LUT ops the executor does not yet run; that is a separate
feature, not affected here.)

The four hw/npu unit tests are unchanged and still pass.

+++ hw/npu/ethos_u_kernels.c
@@ -22,7 +22,7 @@ static inline int8_t sat_i8(int32_t v, int act_min, int act_max)
 }
 
 void ethos_u_conv2d_int8(int8_t *ofm, const int8_t *ifm,
-                         const int8_t *weights, const EthosUScaleBias *sb,
+                         const int16_t *weights, const EthosUScaleBias *sb,
                          const EthosUConvParams *p)
 {
     for (int oy = 0; oy < p->ofm_h; oy++) {
@@ -44,7 +44,7 @@ void ethos_u_conv2d_int8(int8_t *ofm, const int8_t *ifm,
                         }
                         const int8_t *ip = ifm
                             + (iy * p->ifm_w + ix) * p->ifm_c;
-                        const int8_t *wp = weights
+                        const int16_t *wp = weights
                             + ((oc * p->kh + ky) * p->kw + kx) * p->ifm_c;
                         for (int ic = 0; ic < p->ifm_c; ic++) {
                             acc += (int32_t)(ip[ic] - p->ifm_zp) * wp[ic];
@@ -63,7 +63,7 @@ void ethos_u_conv2d_int8(int8_t *ofm, const int8_t *ifm,
 }
 
 void ethos_u_depthwise_int8(int8_t *ofm, const int8_t *ifm,
-                            const int8_t *weights, const EthosUScaleBias *sb,
+                            const int16_t *weights, const EthosUScaleBias *sb,
                             const EthosUConvParams *p)
 {
     for (int oy = 0; oy < p->ofm_h; oy++) {
@@ -84,7 +84,7 @@ void ethos_u_depthwise_int8(int8_t *ofm, const int8_t *ifm,
                             continue;
                         }
                         int8_t in = ifm[(iy * p->ifm_w + ix) * p->ifm_c + c];
-                        int8_t w = weights[(ky * p->kw + kx) * p->ofm_c + c];
+                        int16_t w = weights[(ky * p->kw + kx) * p->ofm_c + c];
                         acc += (int32_t)(in - p->ifm_zp) * w;
                     }
                 }
